@@ -4,11 +4,11 @@ version: 0.1.0
 homepage: https://github.com/xuzhongpeng/git_hooks
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  path: ^1.6.0
-  yaml: ^2.1.0
+  path: ^1.8.0
+  yaml: ^3.0.0
 
 executables:
   git_hooks:
