@@ -58,12 +58,12 @@ class Ansi {
 /// standard status messages, trace level output, and indeterminate progress.
 abstract class Logger {
   /// Create a normal [Logger]; this logger will not display trace level output.
-  factory Logger.standard({Ansi ansi}) => new StandardLogger(ansi: ansi);
+  factory Logger.standard({Ansi? ansi}) => new StandardLogger(ansi: ansi);
 
   /// Create a [Logger] that will display trace level output.
   ///
   /// If [logTime] is `true`, this logger will display the time of the message.
-  factory Logger.verbose({Ansi ansi, bool logTime: true}) {
+  factory Logger.verbose({Ansi? ansi, bool logTime: true}) {
     return new VerboseLogger(ansi: ansi, logTime: logTime);
   }
 
@@ -98,7 +98,7 @@ abstract class Progress {
   Duration get elapsed => _stopwatch.elapsed;
 
   /// Finish the indeterminate progress display.
-  void finish({String message, bool showTiming});
+  void finish({String? message, bool showTiming = false});
 
   /// Cancel the indeterminate progress display.
   void cancel();
@@ -107,17 +107,16 @@ abstract class Progress {
 class StandardLogger implements Logger {
   Ansi ansi;
 
-  StandardLogger({this.ansi}) {
-    ansi ??= new Ansi(Ansi.terminalSupportsAnsi);
-  }
+  StandardLogger({Ansi? ansi})
+      : this.ansi = ansi ?? new Ansi(Ansi.terminalSupportsAnsi);
 
   bool get isVerbose => false;
 
-  Progress _currentProgress;
+  Progress? _currentProgress;
 
   void stderr(String message) {
-    if (_currentProgress != null) {
-      Progress progress = _currentProgress;
+    if (_currentProgress is Progress) {
+      Progress progress = _currentProgress!;
       _currentProgress = null;
       progress.cancel();
     }
@@ -126,8 +125,8 @@ class StandardLogger implements Logger {
   }
 
   void stdout(String message) {
-    if (_currentProgress != null) {
-      Progress progress = _currentProgress;
+    if (_currentProgress is Progress) {
+      Progress progress = _currentProgress!;
       _currentProgress = null;
       progress.cancel();
     }
@@ -138,8 +137,8 @@ class StandardLogger implements Logger {
   void trace(String message) {}
 
   Progress progress(String message) {
-    if (_currentProgress != null) {
-      Progress progress = _currentProgress;
+    if (_currentProgress is Progress) {
+      Progress progress = _currentProgress!;
       _currentProgress = null;
       progress.cancel();
     }
@@ -166,7 +165,7 @@ class SimpleProgress extends Progress {
   void cancel() {}
 
   @override
-  void finish({String message, bool showTiming}) {}
+  void finish({String? message, bool showTiming = false}) {}
 }
 
 class AnsiProgress extends Progress {
@@ -175,7 +174,7 @@ class AnsiProgress extends Progress {
   final Ansi ansi;
 
   int _index = 0;
-  Timer _timer;
+  late Timer _timer;
 
   AnsiProgress(this.ansi, String message) : super(message) {
     io.stdout.write('${message}...  '.padRight(40));
@@ -197,18 +196,19 @@ class AnsiProgress extends Progress {
   }
 
   @override
-  void finish({String message, bool showTiming: false}) {
+  void finish({String? message, bool showTiming = false}) {
     if (_timer.isActive) {
       _timer.cancel();
       _updateDisplay(isFinal: true, message: message, showTiming: showTiming);
     }
   }
 
-  void _updateDisplay(
-      {bool isFinal: false,
-      bool cancelled: false,
-      String message,
-      bool showTiming: false}) {
+  void _updateDisplay({
+    bool isFinal = false,
+    bool cancelled = false,
+    String? message,
+    bool showTiming = false,
+  }) {
     String char = kAnimationItems[_index % kAnimationItems.length];
     if (isFinal || cancelled) {
       char = '';
@@ -233,12 +233,11 @@ class VerboseLogger implements Logger {
   bool logTime;
   Stopwatch _timer;
 
-  VerboseLogger({this.ansi, this.logTime}) {
-    ansi ??= new Ansi(Ansi.terminalSupportsAnsi);
-    logTime ??= false;
-
-    _timer = new Stopwatch()..start();
-  }
+  VerboseLogger({
+    Ansi? ansi,
+    this.logTime = false,
+  })  : this.ansi = ansi ?? new Ansi(Ansi.terminalSupportsAnsi),
+        _timer = new Stopwatch()..start();
 
   bool get isVerbose => true;
 
