@@ -1,25 +1,25 @@
 import "dart:io";
+
+import 'package:git_hooks/install/create_hooks.dart';
 import 'package:git_hooks/runtime/git_hooks.dart';
 import 'package:git_hooks/utils/utils.dart';
 import "package:yaml/yaml.dart";
 
-import 'package:git_hooks/install/create_hooks.dart';
-
 main(List<String> arguments) {
   if (arguments.isNotEmpty) {
     String str = arguments[0];
-    if (arguments != null && arguments.length >= 1) {
+    if (arguments.length >= 1) {
       if (str == 'create') {
         //init files
         String targetPath = arguments[1];
-        if (targetPath != null && targetPath.endsWith(".dart")) {
+        if (targetPath is String && targetPath.endsWith(".dart")) {
           CreateHooks.copyFile(targetPath: targetPath);
         } else
           CreateHooks.copyFile();
       } else if (str == '-h' || str == '-help') {
         help();
       } else if (str == '-v' || str == '--version') {
-        File f = new File(Utils.uri(Utils.getOwnPath() + "/pubspec.yaml"));
+        File f = new File(Utils.uri(Utils.getOwnPath()! + "/pubspec.yaml"));
         String text = f.readAsStringSync();
         Map yaml = loadYaml(text);
         String version = yaml['version'];
@@ -31,8 +31,7 @@ main(List<String> arguments) {
         print('');
         help();
       }
-    }
-    else {
+    } else {
       print(
           "Too many positional arguments: 1 expected, but ${arguments.length} found");
       print('');
